Add a Go version of update.sh.
The configfiles updater can be run as a daemon, in which case it will
update configfiles every 1m after connectivity changes (more likely to
succeed), but also every hour.

It can still be triggered from the shell, and defaults to that behavior.

To build, use `GOPATH=$PWD/gocode go install configfiles/...`